# SSE2 optimization of addWeighted