```
addWeighted::Size_MatType::(640x480, 8SC1)              2.257 ms  0.166 ms    13.58   
addWeighted::Size_MatType::(640x480, 16SC1)             2.211 ms  0.162 ms    13.65   
addWeighted::Size_MatType::(1280x720, 8SC1)             6.750 ms  0.493 ms    13.69   
addWeighted::Size_MatType::(1280x720, 16SC1)            6.638 ms  0.448 ms    14.82   
addWeighted::Size_MatType::(1920x1080, 8SC1)            15.111 ms 1.156 ms    13.07   
addWeighted::Size_MatType::(1920x1080, 16SC1)           14.994 ms 1.135 ms    13.21  
```